Separating automatae
